% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_season.R
\name{assign_season}
\alias{assign_season}
\title{Filter phenological cycles}
\usage{
assign_season(
  pheno,
  seasons = NA,
  pop = NULL,
  pop_win = NULL,
  sos_win = NULL,
  eos_win = NULL,
  pop_name = "pop",
  sos_name = "sos",
  eos_name = "eos",
  max_n_cycles = 2,
  rm_unassigned = TRUE
)
}
\arguments{
\item{pheno}{Cycle allocation (data table generated by \code{extract_pheno()}
or \code{cut_cycles()}).}

\item{seasons}{Character vector with the season names to be used.}

\item{pop_name}{Character name of the field in \code{pheno} to be used as metric
for the date of peak (associated with \code{pop_seasons}).}

\item{sos_name}{Character name of the field in \code{pheno} to be used as metric
for the date of start of cycle (associated with \code{sos_seasons}).
Set to NULL in order not to apply.}

\item{eos_name}{Character name of the field in \code{pheno} to be used as metric
for the date of end of cycle (associated with \code{pop_seasons}).
Set to NULL in order not to apply.}

\item{max_n_cycles}{(optional) Maximum number of cycles to be detected in one
season (default: Inf, meaning that all the identified cycles are kept).}

\item{rm_unassigned}{(optional) Logical: should cycles which do not match with
any season (basing on the settings) be dropped from the output?
(Default: TRUE)}

\item{pop_seasons}{List with the allowed ranges for the dates of
cycle peaks (one per season): each element is a character vector with two
elements, in the form \code{"mm-dd"} (month-day), representing the range of the
temporal window (see default value as example) within the corresponding
season. The list must be of the same length of \code{seasons}.
If \code{season} is one-length, \code{pop_seasons} can be both a 1-length list or
a vector.
Each element of the list can be left to NA in order not to specify any
filtering condition.}

\item{sos_seasons}{List with the allowed ranges for the dates of
start of cycle (see \code{pop_seasons} for details about the format).}

\item{eos_seasons}{Named list with the allowed ranges for the dates of
end of cycle (see \code{pop_seasons} for details about the format).}
}
\value{
The input data table, filtered basing on arguments and with the
addition of the field \code{season}, containing the name of each season
(one among the ones specified in argument \code{seasons}) associated to each
cycle.
If \code{seasons = NA} this field is not returned.
}
\description{
Filter detected cycles basing on dates of begin / end / peak,
or limit the number of cycles allowed for any specified season.
}
\author{
Luigi Ranghetti, PhD (2020) \email{luigi@ranghetti.info}
}
