% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cf}
\alias{cf}
\title{Sample curve fitting interpolation}
\format{
A named list of 2 elements (one per input ID) in the output format
of function \code{\link[=fit_curve]{fit_curve()}}.
}
\usage{
data(cf)
}
\description{
Object (list) containing the interpolation (curve fitting) of
some seasonal cycles (\code{\link{dt_cycles}}) extracted from the dataset
\code{\link{ts_filled}}, created following the example provided in the
\code{\link[=fit_curve]{fit_curve()}} documentation.
It is used in the documentation of function \code{\link[=extract_pheno]{extract_pheno()}}.
}
\keyword{datasets}
