% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_cycles.R
\name{cut_cycles}
\alias{cut_cycles}
\title{Cut cycles}
\usage{
cut_cycles(
  ts,
  n_cycles = Inf,
  min_win = 60,
  min_relh = 0.25,
  max_relval = 0.3,
  newyearday = "01-01",
  weight_metric = "integral"
)
}
\arguments{
\item{ts}{Time series in \code{s2ts} format (generated using \code{fill_s2ts()}).}

\item{n_cycles}{(optional) Maximum number of cycles to be detected in one year
(default: Inf, meaning that all the identified cycles are kept).
A cycle overlapping the new year's day (argument \code{newyearday})
is assigned to the year in which the date of maximum value falls.}

\item{min_win}{(optional) Minimum time window between two consecutive
maxima / minima to consider a separate cycle.}

\item{min_relh}{(optional) Minimum relative difference between the maximum
and each of the two minima to consider a separate cycle.}

\item{max_relval}{(optional) Maximum relative value to consider a
cycle breakpoint.}

\item{newyearday}{(optional) day to be considered as new year's day, used
to assign cycles to the proper year.
It can be an object of class \code{Date} (in which case the year is ignored)
or a character value in the form \code{mm-dd}.
In case it is July 1 or higher, cycles whose maximum value is falling in
the last  part of the year are assigned to the subsequent year;
otherwise, cycles whose maximum value is falling in the first part of the
year are assigned  to the previous year. Default is January 1
(all cycles are assigned the the year in which their maximum value falls).}

\item{rank_metric}{(optional) TODO}
}
\value{
A data table with the following fields:
\itemize{
\item \code{id}: the time series ID (see \code{s2ts});
\item \code{year}: the year assigned to each cycle;
\item \code{cycle}: the cycle ID (progressive integer within each year);
\item \code{begin}: the date of the begin of the cycle;
\item \code{end}: the date of the end of the cycle;
\item \code{maxval}: the date of the maximum value of the cycle;
\item \code{weight}: the value of the metric used for ranking seasons.
}
}
\description{
Cut Sentinel-2 time series into separate cycles,
detecting dates of cuts and of peaks.
}
\note{
The steps used to discriminate seasons are partially based on the method
exposed in Meroni et al. (2021) (\doi{10.1016/j.rse.2020.112232}).
The methodology will be documented in future.
}
\author{
Luigi Ranghetti, PhD (2021) \email{luigi@ranghetti.info}
}
