% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_seasons.R
\name{cut_seasons}
\alias{cut_seasons}
\title{Cut seasons}
\usage{
cut_seasons(ts, min_win = 60, min_relh = 0.25, max_relval = 0.3)
}
\arguments{
\item{ts}{Time series in \code{s2ts} format (generated using \code{fill_s2ts()}).}

\item{min_win}{(optional) Minimum time window between two consecutive
maxima / minima to consider a separate season.}

\item{min_relh}{(optional) Minimum relative difference between the maximum
and each of the two minima to consider a separate season.}

\item{max_relval}{(optional) Maximum relative value to consider a
season breakpoint.}
}
\value{
A data table with the following fields:
\itemize{
\item \code{id}: the time series ID (see \code{s2ts});
\item \code{season}: the season ID (integer);
\item \code{begin}: the date of the begin of the season;
\item \code{end}: the date of the end of the season;
\item \code{maxval}: the date of the maximum value of the season.
}
}
\description{
Cut Sentinel-2 time series into separate seasons,
detecting dates of cuts and of peaks.
}
\author{
Luigi Ranghetti, PhD (2020) \email{luigi@ranghetti.info}
}
