% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dt_cycles}
\alias{dt_cycles}
\title{Sample seasonal cycles}
\format{
A data table with 3 records (identified cycles) and the following
fields:
\itemize{
\item \code{id}: the time series ID;
\item \code{year}: the year assigned to each cycle;
\item \code{cycle}: the cycle ID (progressive integer within each year);
\item \code{begin}: the date of the begin of the cycle;
\item \code{end}: the date of the end of the cycle;
\item \code{maxval}: the date of the maximum value of the cycle;
\item \code{weight}: the value of the metric used for ranking seasons.
}
}
\usage{
data(dt_cycles)
}
\description{
Data frame with the cycles extracted from the dataset
\code{\link{ts_filled}}, created following the example provided in the
\code{\link[=cut_cycles]{cut_cycles()}} documentation.
It is used in the documentation of function \code{\link[=fit_curve]{fit_curve()}}.
}
\examples{
data(dt_cycles)
dt_cycles
}
\keyword{datasets}
