% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dt_pheno}
\alias{dt_pheno}
\title{Sample phenological metrics dataset}
\format{
A data table with 3 records (identified cycles) and the following
fields:
\itemize{
\item \code{id}, \code{year}, \code{cycle}, \code{begin}, \code{end}, \code{maxval}, \code{weight}:
inherited from \code{\link{dt_cycles}}
\item \code{sos}, \code{eos}, \code{los}, \code{pop}, \code{mgs}, \code{rsp}, \code{rau}, \code{peak}, \code{msp}, \code{mau}:
phenological metrics created by method \code{"trs"} (see
\code{phenopix::PhenoTrs()}).
}
}
\usage{
data(dt_pheno)
}
\description{
Data frame with the phenological metrics extracted from the
object \code{\link{cf}}, created following the example provided in the
\code{\link[=extract_pheno]{extract_pheno()}} documentation.
It is used in the documentation of functions \code{\link[=assign_season]{assign_season()}} and
\code{\link[=aggregate_pheno]{aggregate_pheno()}}.
}
\examples{
data(dt_pheno)
dt_pheno
}
\keyword{datasets}
