% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_mowing.R
\name{extract_mowing}
\alias{extract_mowing}
\title{Extract mowing dates}
\usage{
extract_mowing(
  ts,
  cycles,
  min_diff = 0.5,
  max_post = 0.3,
  max_length = Inf,
  max_n_mow = 3,
  max_d_mow = 10,
  value_type = "senescence",
  min_qa = 0.6
)
}
\arguments{
\item{ts}{Time series in \code{s2ts} format (generated using \code{extract_s2ts()}
or \code{smooth_s2ts()} -- \strong{not} \code{fill_s2ts()}).}

\item{cycles}{Cycles allocation (data table generated by \code{\link[=cut_cycles]{cut_cycles()}},
eventually filtered using \code{\link[=assign_season]{assign_season()}}).}

\item{min_diff}{minimum difference (absolute or relative, depending on
argument \code{value_type}) among pre- and post-event values to discriminate
a mowing window (default: 0.5).}

\item{max_post}{maximum post-event value (absolute or relative, depending on
argument \code{value_type}) to discriminate a mowing window (default: 0.3).}

\item{max_length}{maximum number of days passing among pre- and post-event
dates to discriminate a mowing window (default: no limit).}

\item{max_n_mow}{maximum allowed number of images among pre- and post-event
dates (default: 3).}

\item{max_d_mow}{maximum number of days between an intermediate image and
a pre- or post-event date (default: 10).}

\item{value_type}{(optional) Character:
\itemize{
\item if \code{"absolute"}, value set with argument \code{min_diff} is an absolute value;
\item if \code{"senescence"} (default) value set with argument \code{min_diff} is a
a relative measure among the maximum (peak) and the minimum (after drop);
\item if \code{"cycle"} value set with argument \code{min_diff} is a
a relative measure among the range of the specific cycle;
\item if \code{"season"} value set with argument \code{min_diff} is a
a relative measure among the range of the crop season;
}}

\item{min_qa}{(optional) minimum 0-1 quality value
(points with \code{qa < min_qa} are not used, while \code{qa} values
in the range \code{min_qa} to 1 are reshaped into the range 0 to 1).
Default is 0.5. It only applies if \code{ts} is a raw TS.}
}
\value{
A data table with the following fields:
\itemize{
\item \code{id}: the time series ID (see \code{s2ts});
\item \code{year}: the year (integer);
\item \code{cycle}: the cycle ID (integer);
\item \code{begin}, \code{end}, \code{maxval}: the dates of begin, end and maximum value in
the cycle as computed by \code{cut_cycles()};
\item other columns included in input \code{cycle} dt (e.g. \code{"weight"});
\item \code{mowing}: estimated mowing date;
\item \code{mowing_start}: estimated date of beginning of mowing phase;
\item \code{mowing_end}: estimated date of ending of mowing phase.
}
}
\description{
Extract mowing dates from a raw or smoothed time series.
}
\examples{
# Load input data
data("ts_raw")
data("dt_cycles")
data("ts_filled") # used for plots

# Default extraction ("trs" method with 50\% threshold)
mowing_dt <- extract_mowing(ts_raw, dt_cycles) 
plot(ts_filled, pheno = dt_cycles, plot_points = TRUE) +
  ggplot2::geom_vline(ggplot2::aes(xintercept=mowing), mowing_dt, linetype="dashed")
}
\author{
Luigi Ranghetti, PhD (2024)
}
