% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_pheno.R
\name{extract_pheno}
\alias{extract_pheno}
\title{Extract phenological metrics}
\usage{
extract_pheno(data, method = "trs", trs = 0.5, ...)
}
\arguments{
\item{data}{List of fitted time series as generated by function \code{fit_curve()}.}

\item{method}{Thresholding method among \code{"trs"}, \code{"derivatives"},
\code{"klosterman"} and \code{"gu"} (see \code{phenopix::PhenoExtract()}).}

\item{trs}{Argument passed to \code{phenopix::PhenoExtract()} if \code{method = "trs"}.}

\item{...}{Additional arguments passed to \code{phenopix::PhenoExtract()}.}
}
\value{
A data table with the following fields:
\itemize{
\item \code{id}: the time series ID (see \code{s2ts});
\item \code{season}: the season ID (integer);
\item \code{begin}, \code{end}, \code{maxval}: the dates of begin, end and maximum value in
the season as computed by \code{cut_seasons()};
\item other phenology metrics, depending on \code{method}:
\itemize{
\item if \code{method = "trs"} or \code{"derivatives"}: \code{sos}, \code{eos}, \code{los}, \code{pop},
\code{mgs}, \code{rsp}, \code{rau}, \code{peak}, \code{msp}, \code{mau} (see
\code{phenopix::PhenoTrs()} or \code{phenopix::PhenoDeriv()});
\item if \code{method = "gu"}: \code{UD}, \code{SD}, \code{DD}, \code{RD}, \code{maxline}, \code{baseline},
\code{prr}, \code{psr}, \code{plateau.slope} (see \code{phenopix::PhenoGu()});
\item if \code{method = "klosterman"}: \code{Greenup}, \code{Maturity}, \code{Senescence},
\code{Dormancy} (see \code{phenopix::PhenoKl()}).
}
}
}
\description{
Extract phenological metrics from a fitted time series.
}
\author{
Luigi Ranghetti, PhD (2020) \email{luigi@ranghetti.info}
}
