% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_s2ts.R
\name{extract_s2ts}
\alias{extract_s2ts}
\title{Extract time series from sen2r archives}
\usage{
extract_s2ts(
  in_paths,
  in_sf,
  fun = mean,
  in_sf_id = NA,
  scl_paths = NULL,
  scl_weights = NA
)
}
\arguments{
\item{in_paths}{Paths of the sen2r files (eventually obtained using
\code{read_in_cube(..., out_format = "path")}).}

\item{in_sf}{Object with polygonal or point geometries}

\item{fun}{(optional) aggregation function to be used in case of polygonal
\code{in_sf} features.
If \code{scl_paths} is defined it is not used for aggregation of output values
(\code{weighted.mean()} is always used for that), but for aggregation of weights
made to compute the \code{"quality"} attribute.}

\item{in_sf_id}{(optional) charachter vector corresponding to the name/names
of the \code{in_sf} column with the features IDs.
If NA (default) the row number is used.}

\item{scl_paths}{(optional) Paths of the SCL files (they must correspond
to \code{in_paths}); if provided, it is used to weight pixels
during the aggregation and to provide an output quality flag.}

\item{scl_weights}{(optional) weights to be used for each SCL class,
which can be created using function \code{scl_weights()}.}

\item{min_cov}{(optional) quality threshold (0-1) for output values to be
returned (default is 0, meaning that all values are returned).}
}
\value{
The output time series in tabular format, containing the following
columns:
\itemize{
\item \code{date} :date of observation;
\item \code{id} : polygon ID, corresponding to argument \code{in_sf_id};
\item \code{value}: extracted value;
\item \code{quality}: relative 0-1 quality values (missing if \code{scl_paths} was not provided).
}
}
\description{
TODO
}
\author{
Luigi Ranghetti, phD (2020) \email{luigi@ranghetti.info}
}
