% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_curve.R
\name{fakeFit}
\alias{fakeFit}
\title{Covert a \code{zoo} into a \code{{phenopix}} format}
\usage{
fakeFit(ts, uncert = NA, nrep = NA)
}
\arguments{
\item{ts}{A \code{ts} or \code{zoo} object with \code{gcc} data.
index (\code{ts}) must be numeric days of year (\code{DOY}) or a \code{POSIXct} vector}

\item{uncert}{Unused (left for compatibility).}

\item{nrep}{Unused (left for compatibility).}
}
\value{
A list containing the following items.
\itemize{
\item \code{fit}: A list as returned by the function \code{FitDoubleLogElmore}
\item \code{uncertainty}: A list containing a \code{zoo} \code{data.frame} with the
uncertainty predicted values, and a dataframe containing the
respective uncertainty curve parameters
}
}
\description{
Internal function to convert a \code{zoo} time series in the format
returned by functions like \code{phenopix::ElmoreFit()} or \code{phenopix::GuFit()}.
}
\author{
Luigi Ranghetti, PhD (2020) \email{luigi@ranghetti.info}
}
