% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_seasons.R
\name{filter_seasons}
\alias{filter_seasons}
\title{Filter growing seasons}
\usage{
filter_seasons(
  seas,
  pop_win = list(winter = c("02-01", "05-31"), summer = c("06-01", "09-30")),
  sos_win = list(winter = c("10-01", "04-30"), summer = c("05-01", "08-31")),
  eos_win = list(winter = c("03-01", "07-15"), summer = c("07-16", "11-15")),
  pop_metric = "pop",
  sos_metric = "sos",
  eos_metric = "eos",
  peak_metric = "peak",
  newyear = "10-01"
)
}
\arguments{
\item{seas}{Seasons allocation (data table generated by \code{extract_pheno()}
or \code{cut_seasons()}).}

\item{pop_metric}{Character name of the field in \code{seas} to be used as metric
for the date of peak (associated with \code{pop_seasons}).}

\item{sos_metric}{Character name of the field in \code{seas} to be used as metric
for the date of beginning of growing season (associated with \code{sos_seasons}).}

\item{eos_metric}{Character name of the field in \code{seas} to be used as metric
for the date of end of growing season (associated with \code{pop_seasons}).}

\item{peak_metric}{Character name of the field in \code{seas} to be used as metric
for the value of the peak.}

\item{newyear}{Character vector in the form \code{"mm-dd"} (month-day)
representing the day from which a new year is considered.
If it is in July or later, seasons with a peak after day \code{newyear} are
associated with the subsequent year; otherwise, seasons with a peak before
day \code{newyear} are associated with the previous year.}

\item{pop_seasons}{Named list with the allowed ranges for the dates of
seasonal peaks: each element is a character vector with two
elements, in the form \code{"mm-dd"} (month-day), representing the range of the
temporal window (see default value as example).
The number of elements is the maximum number of seasons accepted for each
year ( in case of a higher number, the season with a higher \code{peak} value
within each season name is maintained).}

\item{sos_seasons}{Named list with the allowed ranges for the dates of
beginning of growing season: each element is a character vector with two
elements, in the form \code{"mm-dd"} (month-day), representing the range of the
temporal window (see default value as example).
If must be of the same length of \code{pop_seasons} and with the same names.}

\item{eos_seasons}{Named list with the allowed ranges for the dates of
end of growing season: each element is a character vector with two
elements, in the form \code{"mm-dd"} (month-day), representing the range of the
temporal window (see default value as example).
If must be of the same length of \code{pop_seasons} and with the same names.}
}
\value{
The input data table, filtered basing on arguments and with two
more fields:
\itemize{
\item \code{seas_name}: the name of the season (corresponding to the name of one of
the elements of argument \code{pop_seasons});
\item \code{year}: the year associated to each season.
}
}
\description{
Filter detected seasons basing on dates of begin / end / peak,
and associate each season with a reference year.
}
\author{
Luigi Ranghetti, PhD (2020) \email{luigi@ranghetti.info}
}
