% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_s2paths.R
\name{load_s2paths}
\alias{load_s2paths}
\title{Load raster paths or data cube from a sen2r archive}
\usage{
load_s2paths(
  inpath,
  out_format = "path",
  prod_type,
  time_window,
  s2_orbits,
  s2_sensors,
  file_ext,
  bbox,
  naming_convention = "sen2r",
  .use_vrt = FALSE
)
}
\arguments{
\item{inpath}{Path of the directory in which \code{{sen2r}} files are stored.}

\item{out_format}{(optional) Output format, being one among
\code{"path"} (default), \code{"stars"} and \code{"stars_proxy"}).}

\item{prod_type}{(optional) \code{sen2r} product type to import
(used for filtering among \code{inpath} content).}

\item{time_window}{(optional) time window to import
(used for filtering among \code{inpath} content).}

\item{s2_orbits}{(optional) Sentinel-2 orbits to import
(used for filtering among \code{inpath} content).}

\item{s2_sensors}{(optional) Sentinel-2 sensor to import  (\code{"2A"}, \code{"2B"} or both)
(used for filtering among \code{inpath} content).}

\item{file_ext}{(optional) input file extension
(used for filtering among \code{inpath} content).}

\item{bbox}{(optional) object of class \code{sf}, \code{sfc} or \code{bbox} used to import
a subset of the original raster extent.}

\item{naming_convention}{(optional) the naming convention used to extract
information from file names contained in \code{inpath}
(see \code{sen2r::sen2r_getElements()} for details).
It takes effect only if sen2r version is > 1.5.0.}

\item{.use_vrt}{(optional) pass through a temporary VRT before creating \code{stars}?}
}
\value{
The output raster cube in the selected format.
}
\description{
Load the paths of rasters included in a Sentinel-2 archive
generated with \code{{sen2r}}, filtering relevant files and ordering them.
Output can also be returned in format \code{stars} or \code{stars_proxy}
(see argument \code{out_format}).
}
\examples{
# Path of a sample archive created with sen2r (see ?sample_paths)
archive_dir <- system.file("extdata/sen2r/sampleroi", package = "sen2rts")

# Default function behaviour
sen2r_ndvi_paths <- load_s2paths(file.path(archive_dir, "NDVI"))
gsub(archive_dir, "", head(sen2r_ndvi_paths))

# Specifying argument prod_type, the parent directory can also be provided:
sen2r_scl_paths <- load_s2paths(archive_dir, prod_type = "SCL")
gsub(archive_dir, "", head(sen2r_scl_paths))

# Filter on dates and return a stars raster cube
sen2r_ndvi_cube <- load_s2paths(
  archive_dir, prod_type = "NDVI", 
  time_window = c("2020-05-20", "2020-06-05"), 
  out_format = "stars"
)
plot(sen2r_ndvi_cube)
}
\author{
Luigi Ranghetti, PhD (2020) \email{luigi@ranghetti.info}
}
