% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2ts-class.R
\name{plot}
\alias{plot}
\alias{plot.s2ts}
\title{Plot \code{s2ts} object}
\usage{
\method{plot}{s2ts}(
  x,
  pheno,
  fitted,
  plot_points,
  plot_rawline,
  plot_smoothed,
  plot_fitted,
  plot_cuts,
  plot_cycles = TRUE,
  plot_dates = FALSE,
  pheno_metrics,
  ...
)
}
\arguments{
\item{x}{Object of class \code{s2ts} to be plotted.}

\item{pheno}{(optional) Output of \code{cut_cycles()}}

\item{fitted}{(optional) Output of \code{fit_curve()}}

\item{plot_points}{(optional) Logical: should raw point values be plotted?
Default: only if \code{pheno} and \code{fitted} are not provided.}

\item{plot_rawline}{(optional) Logical: should lines connecting raw points
be plotted? Default: yes.
They are represented as dark grey lines, or as almost transparent lines
if smoothed values exist and are represented.}

\item{plot_smoothed}{(optional) Logical: should lines connecting smoothed
values be plotted? Default: yes (if exist).
They are represented as dark grey lines.}

\item{plot_fitted}{(optional) Logical: should double logistic curves be
plotted? Default: yes, if provided. They are represented as red curves.}

\item{plot_cuts}{(optional) Logical: should cuts between cycles be plotted?
Default: yes, if provided in \code{pheno} or in \code{fitted}.
They are represented as black vertical lines.}

\item{plot_cycles}{(optional) Logical: should existing cycles be highlighted?
If TRUE (default), existing cycles are highlighted with yellow bands.}

\item{plot_dates}{(optional) Logical or character:
if TRUE, plot the dates of cycle cuts and phenology metrics;
if FALSE (default), do not plot anything;
if \code{"cycles"} or \code{"pheno"}, plot only cycle cuts or phenology metrics.}

\item{pheno_metrics}{(optional) Character vector containing the names of
the phenological metrics to be plotted.
If not provided or if \code{pheno_metrics = "all"}, all available metrics
are plotted.}

\item{...}{Not currently used.}
}
\description{
Plot a \code{s2ts} time series, using \code{{ggplot2}} routines.
}
\author{
Luigi Ranghetti, PhD (2020) \email{luigi@ranghetti.info}
}
