% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_sen2r.R
\name{raster_metadata}
\alias{raster_metadata}
\title{Get metadata from raster paths (reduced version)}
\usage{
raster_metadata(raster_paths, meta = "all", format = "list")
}
\arguments{
\item{raster_paths}{A vector of raster paths.}

\item{meta}{Kept for signature compatibility; ignored (metadata
\code{res}, \code{bbox} and \code{proj} are always returned).}

\item{format}{Only \code{"list"} is supported.}
}
\value{
A list of the output metadata.
}
\description{
This accessory function extracts some useful metadata from
a vector of raster paths. This is a reduced version of the homonymous
\code{sen2r::raster_metadata()}, limited to the needs of \code{{sen2rts}}: it only
supports \code{format = "list"} and returns the metadata \code{res}, \code{bbox} and
\code{proj} (plus \code{path} and \code{valid}).
}
\note{
Originally from package {sen2r} (GPL-3); reduced for {sen2rts}.
}
\author{
Luigi Ranghetti, PhD (2019)
}
\keyword{internal}
