% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_s2cube.R
\name{read_s2cube}
\alias{read_s2cube}
\title{Read raster cubes from sen2r archives}
\usage{
read_s2cube(
  inpath,
  out_format = "path",
  prod_type = NA,
  time_window = NA,
  s2_orbits = NA,
  s2_sensors = c("2A", "2B"),
  file_ext = NA,
  bbox = NA,
  .use_vrt = FALSE
)
}
\arguments{
\item{inpath}{Path of the directory in which sen2r files are.}

\item{out_format}{(optional) Output format, being one among
\code{"stars"}, \code{"stars_proxy"} and \code{"path"} -- default is \code{"path"}).}

\item{prod_type}{(optional) sen2r product type to import
(used for filtering among \code{inpath} content).}

\item{time_window}{(optional) time window to import
(used for filtering among \code{inpath} content).}

\item{s2_orbits}{(optional) Sentinel-2 orbits to import
(used for filtering among \code{inpath} content).}

\item{s2_sensors}{(optional) Sentinel-2 sensor to import  (\code{"2A"}, \code{"2B"} or both)
(used for filtering among \code{inpath} content).}

\item{file_ext}{(optional) input file extension
(used for filtering among \code{inpath} content).}

\item{bbox}{(optional) object of class \code{sf}, \code{sfc} or \code{bbox} used to import
a subset of the original raster extent.}

\item{.use_vrt}{(optional) pass through a temporary VRT before creating \code{stars}?}
}
\value{
The output raster cube in the selected format.
}
\description{
TODO
}
\examples{
json_path <- sen2r::build_example_param_file()
out_paths_2 <- sen2r::sen2r(json_path, timewindow = c("2019-07-13", "2019-07-23"))
inpath <- file.path(unique(dirname(dirname(out_paths_2))), "MSAVI2")
extent <- sf::st_read(system.file("extdata/vector/barbellino.geojson", package = "sen2r"))
in_paths <- read_s2cube(inpath)
in_stars <- read_s2cube(inpath, out_format = "stars")
}
\author{
Luigi Ranghetti, PhD (2020) \email{luigi@ranghetti.info}
}
