% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{samplecrops}
\alias{samplecrops}
\title{Boundaries of the sample crops used in the vignette
"The sen2rts workflow"}
\format{
A polygon \code{sf} dataset in UTM 32N CRS
with 12 rows (sample crops) and the variables \code{fid} with crop IDs
("01" to "12").
}
\usage{
data(samplecrops)
}
\description{
Dataset containing boundaries of some field crops of
the farm \href{https://bonificheferraresi.it/it/home}{\emph{Bonifiche Ferraresi}}
located in the estate of Jolanda di Savoia (FE, Italy),
which are used in the vignette
\href{https://sen2rts.ranghetti.info/articles/workflow.html}{"The sen2rts workflow"}.
}
\examples{
data(samplecrops)
samplecrops
plot(samplecrops)
}
\keyword{datasets}
