% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sampleroi}
\alias{sampleroi}
\title{Sample areas used in documentation examples}
\format{
A polygon \code{sf} dataset in UTM 32N CRS
with 2 rows (ROI) and the variables \code{id} with ROI IDs ("01" and "02").
}
\usage{
data(sampleroi)
}
\description{
Dataset containing two small ROI (Regions Of Interests) within
field crops "03" and "04" of the dataset \code{\link{samplecrops}}.
The are used in the documentation examples of \code{\link[=extract_s2ts]{extract_s2ts()}}.
}
\examples{
data(sampleroi)
sampleroi
plot(sampleroi)
}
\keyword{datasets}
