% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scl_weights.R
\name{scl_weights}
\alias{scl_weights}
\title{Build SCL weights}
\usage{
scl_weights(
  no_data = 0,
  saturated_or_defective = 0,
  dark_area_pixels = 0.33,
  cloud_shadows = 0.17,
  vegetation = 1,
  not_vegetated = 1,
  water = 1,
  unclassified = 0.33,
  cloud_medium_probability = 0,
  cloud_high_probability = 0,
  thin_cirrus = 0.33,
  snow = 1
)
}
\arguments{
\item{saturated_or_defective}{(optional) Relative weight (0-1) for SCL class
\code{"saturated_or_defective"}.}

\item{dark_area_pixels}{(optional) Relative weight (0-1) for SCL class
\code{"dark_area_pixels"}.}

\item{cloud_shadows}{(optional) Relative weight (0-1) for SCL class
\code{"cloud_shadows"}.}

\item{vegetation}{(optional) Relative weight (0-1) for SCL class
\code{"vegetation"}.}

\item{not_vegetated}{(optional) Relative weight (0-1) for SCL class
\code{"not_vegetated"}.}

\item{water}{(optional) Relative weight (0-1) for SCL class
\code{"water"}.}

\item{unclassified}{(optional) Relative weight (0-1) for SCL class
\code{"unclassified"}.}

\item{cloud_medium_probability}{(optional) Relative weight (0-1) for SCL class
\code{"cloud_medium_probability"}.}

\item{cloud_high_probability}{(optional) Relative weight (0-1) for SCL class
\code{"cloud_high_probability"}.}

\item{thin_cirrus}{(optional) Relative weight (0-1) for SCL class
\code{"thin_cirrus"}.}

\item{snow}{(optional) Relative weight (0-1) for SCL class
\code{"snow"}.}

\item{nodata}{(optional) Relative weight (0-1) for SCL class
\code{"nodata"}.}
}
\value{
A named vector of length 12, containing 0-1 relative weights
for each SCL class.
}
\description{
Build a vector with relative weights for each SCL class.
Class weights can be defined using function arguments  (all of them
are optional; default weights are used for each non-provided class).
}
\examples{
# Default weights
scl_weights()
# Custom weights changing two classes
scl_weights(unclassified = 0.25, snow = 0)
}
\author{
Luigi Ranghetti, PhD (2020) \email{luigi@ranghetti.info}
}
