% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_sen2r.R
\name{sen2r_getElements}
\alias{sen2r_getElements}
\title{Get information from S2 short name}
\usage{
sen2r_getElements(
  s2_names,
  naming_convention,
  format = "data.table",
  abort = TRUE
)
}
\arguments{
\item{s2_names}{A vector of Sentinel-2 product names in the
sen2r naming convention.}

\item{naming_convention}{The naming convention used to extract information
from \code{s2_names} names (\code{"sen2r"} or \code{"sen2r_new"}, or a custom list).}

\item{format}{One between \code{data.table} (default), \code{data.frame} and \code{list}.}

\item{abort}{Logical parameter: if TRUE (default), the function aborts
in case any of \code{s2_names} is not recognised; if FALSE, a warning is shown.}
}
\value{
A data.table, data.frame or list of the output metadata.
}
\description{
This accessory function extracts metadata included in
the name of a Sentinel-2 product which follows the sen2r naming convention.
}
\note{
Originally from package sen2r (GPL-3).
}
\author{
Luigi Ranghetti, PhD (2019)
}
\keyword{internal}
