% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ts_filled}
\alias{ts_filled}
\title{Sample daily time series}
\format{
A \code{\link{s2ts}} object with 301 dates (in the period 2020-01-01 to
2020-09-30) and 2 IDs (corresponding to \code{\link{sampleroi}} polygons).
}
\usage{
data(ts_filled)
}
\description{
Sample time series obtained from the dataset \code{\link{ts_smoothed}},
created following the example provided in the \code{\link[=fill_s2ts]{fill_s2ts()}}
documentation.
It is used in the documentation of functions with takes a processed
time series as input, like \code{\link[=cut_cycles]{cut_cycles()}}, \code{\link[=fit_curve]{fit_curve()}} and
\code{\link[=aggregate_pheno]{aggregate_pheno()}}.
}
\examples{
data(ts_filled)
print(ts_filled, topn = 5) # standard print
head(as.data.frame(ts_filled)) # see content
plot(ts_filled)
}
\keyword{datasets}
