% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ts_raw}
\alias{ts_raw}
\title{Sample raw time series}
\format{
A \code{\link{s2ts}} object with 60 dates (in the period 2020-01-01 to
2020-09-30) and 2 IDs (corresponding to \code{\link{sampleroi}} polygons).
}
\usage{
data(ts_raw)
}
\description{
Sample time series extracted from the example image archive
provided with the package (see \code{\link[=sample_paths]{sample_paths()}})
over the extension of \code{\link{sampleroi}} polygons.
It was created following the example provided in the \code{\link[=extract_s2ts]{extract_s2ts()}}
documentation.
It is used in the documentation of function \code{\link[=smooth_s2ts]{smooth_s2ts()}}.
}
\examples{
data(ts_raw)
print(ts_raw, topn = 5) # standard print
head(as.data.frame(ts_raw)) # see content
plot(ts_raw)
}
\keyword{datasets}
