% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ts_smoothed}
\alias{ts_smoothed}
\title{Sample smoothed time series}
\format{
A \code{\link{s2ts}} object with 60 dates (in the period 2020-01-01 to
2020-09-30) and 2 IDs (corresponding to \code{\link{sampleroi}} polygons).
}
\usage{
data(ts_smoothed)
}
\description{
Sample time series obtained from the dataset \code{\link{ts_raw}},
created following the example provided in the \code{\link[=smooth_s2ts]{smooth_s2ts()}}
documentation.
It is used in the documentation of function \code{\link[=fill_s2ts]{fill_s2ts()}}.
}
\examples{
data(ts_smoothed)
print(ts_smoothed, topn = 5) # standard print
head(as.data.frame(ts_smoothed)) # see content
plot(ts_smoothed)
}
\keyword{datasets}
