% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w_savgol.R
\name{w_savgol}
\alias{w_savgol}
\title{Savitzky-Golay smoothing filter for not equally spaced data}
\usage{
w_savgol(x, y, q, window, polynom)
}
\arguments{
\item{x}{List of floats representing the "x" values of the data.}

\item{y}{List of floats representing the "y" values. Must have same length as \code{x}.}

\item{q}{List of floats representing the relative weight of "y" values. Must have same length as \code{y}.}

\item{window}{Window length of datapoints. Must be odd and smaller than \code{x}.}

\item{polynom}{The order of polynom used. Must be smaller than the \code{window} size.}
}
\value{
The smoothed "y" values.
}
\description{
TODO
}
\note{
This is an R adaptation of Python function at
https://dsp.stackexchange.com/questions/1676/savitzky-golay-smoothing-filter-for-not-equally-spaced-data
}
\author{
Luigi Ranghetti, phD (2020) \email{luigi@ranghetti.info}
}
