% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_seasons.R
\name{cut_seasons}
\alias{cut_seasons}
\title{Cut seasons}
\usage{
cut_seasons(ts, min_win = 60, min_relh = 0.25, max_relval = 0.3)
}
\arguments{
\item{ts}{Time series in \code{s2ts} format (generated using \code{fill_s2ts()}).}

\item{min_win}{(optional) Minimum time window between two consecutive
maxima / minima to consider a separate season.}

\item{min_relh}{(optional) Minimum relative difference between the maximum
and each of the two minima to consider a separate season.}

\item{max_relval}{(optional) Maximum relative value to consider a
season breakpoint.}
}
\value{
The output time series in tabular format (see \code{extract_ts()}),
with a new column \code{pheno} in which records detected as season breakpoints
are marked with \code{"cut_seas"}, while season peaks with \code{"peak"}.
}
\description{
Cut Sentinel-2 time series into separate seasons,
detecting dates of cuts and of peaks.
}
\author{
Luigi Ranghetti, PhD (2020) \email{luigi@ranghetti.info}
}
