% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_mowing.R
\name{extract_mowing}
\alias{extract_mowing}
\title{Extract phenological metrics}
\usage{
extract_mowing(
  ts,
  cycles,
  min_diff = 0.5,
  max_n_mow = 3,
  max_d_mow = 10,
  value_type = "cycle",
  min_qa = 0.6
)
}
\arguments{
\item{ts}{Time series in \code{s2ts} format (generated using \code{extract_s2ts()} or \code{smooth_s2ts()}).}

\item{cycles}{Cycles allocation (data table generated by \code{\link[=cut_cycles]{cut_cycles()}},
eventually filtered using \code{\link[=assign_season]{assign_season()}}).}

\item{value_type}{(optional) Character:
\itemize{
\item if \code{"absolute"}, value set with argument \code{min_diff} is an absolute value;
\item if \code{"senescence"}, value set with argument \code{min_diff} is a
a relative measure among the maximum (peak) and the minimum (after drop);
\item if \code{"cycle"} (default), value set with argument \code{min_diff} is a
a relative measure among the range of the specific cycle;
\item if \code{"season"}, value set with argument \code{min_diff} is a
a relative measure among the range of the crop season;
}}

\item{min_qa}{(optional) minimum 0-1 quality value
(points with \code{qa < min_qa} are not used, while \code{qa} values
in the range \code{min_qa} to 1 are reshaped into the range 0 to 1).
Default is 0.5. It only applies if \code{ts} is a raw TS.}
}
\value{
A data table with the following fields:
\itemize{
\item \code{id}: the time series ID (see \code{s2ts});
\item \code{year}: the year (integer);
\item \code{cycle}: the cycle ID (integer);
\item \code{begin}, \code{end}, \code{maxval}: the dates of begin, end and maximum value in
the cycle as computed by \code{cut_cycles()};
\item other phenology metrics, depending on \code{method}:
\itemize{
\item if \code{method = "trs"} or \code{"derivatives"}: \code{sos}, \code{eos}, \code{los}, \code{pop},
\code{mgs}, \code{rsp}, \code{rau}, \code{peak}, \code{msp}, \code{mau} (see
\code{phenopix::PhenoTrs()} or \code{phenopix::PhenoDeriv()});
\item if \code{method = "gu"}: \code{UD}, \code{SD}, \code{DD}, \code{RD}, \code{maxline}, \code{baseline},
\code{prr}, \code{psr}, \code{plateau.slope} (see \code{phenopix::PhenoGu()});
\item if \code{method = "klosterman"}: \code{Greenup}, \code{Maturity}, \code{Senescence},
\code{Dormancy} (see \code{phenopix::PhenoKl()}).
}
}
}
\description{
Extract phenological metrics from a fitted time series.
}
\examples{
# Load input data
data("cf")
data("ts_filled") # used for plots

# Default extraction ("trs" method with 50\% threshold)
dt_pheno <- extract_pheno(cf)
plot(ts_filled, fitted = cf, pheno = dt_pheno, plot_dates = TRUE)

\donttest{
# Customize parameters (e.g. "derivatives" method with 30\% threshold)
dt_pheno_2 <- extract_pheno(cf, method = "derivatives", trs = 0.3)
plot(ts_filled, fitted = cf, pheno = dt_pheno_2, plot_dates = TRUE)

# Other methods: Klosterman
dt_pheno_kl <- extract_pheno(cf, method = "klosterman")
plot(ts_filled, fitted = cf, pheno = dt_pheno_kl, plot_dates = TRUE)

# Other methods: Gu
dt_pheno_gu <- extract_pheno(cf, method = "gu")
plot(ts_filled, fitted = cf, pheno = dt_pheno_gu, plot_dates = TRUE)
}
}
\author{
Luigi Ranghetti, PhD (2020) \email{luigi@ranghetti.info}
}
