% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_s2ts.R
\name{extract_s2ts}
\alias{extract_s2ts}
\title{Extract time series from sen2r archives}
\usage{
extract_s2ts(
  s2cube,
  polygons,
  fun = "mean",
  max_cells_in_memory = 3e+07,
  scl_s2cube = NULL,
  scl_weights = NA
)
}
\arguments{
\item{s2cube}{Raster time series in \code{s2cube} format}

\item{fun}{(optional) aggregation function to be used in case of polygonal
\code{in_sf} features (see \code{exactextractr::exact_extract()});
it is ignored in case of point features.}

\item{max_cells_in_memory}{(optional) argument passed to
\code{exactextractr::exact_extract()} (it is ignored in case of point features).}

\item{scl_s2cube}{(optional) Object of class \code{s2cube} containing the SCL maps
for each date in \code{s2cube}; if provided, it is used to weight pixels
during the aggregation and to provide an output quality flag.}

\item{scl_weights}{(optional) weights to be used for each SCL class,
which can be created using function \code{scl_weights()}.}

\item{in_sf}{Object with polygonal or point geometries}

\item{time_window}{(optional) time window to import
(used for filtering among \code{inpath} content).}
}
\value{
The output time series in \code{s2ts} format.
}
\description{
TODO
}
\author{
Luigi Ranghetti, phD (2020) \email{luigi@ranghetti.info}
}
