% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_s2ts.R
\name{fill_s2ts}
\alias{fill_s2ts}
\title{Fill not equally-spaced time series}
\usage{
fill_s2ts(s2ts, frequency = "dop")
}
\arguments{
\item{s2ts}{Time series generated using \code{smooth_s2ts()}.}

\item{frequency}{(optional) One of the followings:
\itemize{
\item \code{dop} (Days Of Passage -- default): values are returned corresponding to
the theoretic Sentinel-2 dates of passage;
\item \code{daily}: daily frequency.
}}
}
\value{
The output time series in tabular format (see \code{extract_ts()}).
}
\description{
TODO
}
\author{
Luigi Ranghetti, phD (2020) \email{luigi@ranghetti.info}
}
