% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_s2ts.R
\name{fill_s2ts}
\alias{fill_s2ts}
\title{Fill not equally-spaced time series}
\usage{
fill_s2ts(
  ts,
  frequency = "daily",
  method = "fmm",
  max_na_days = Inf,
  max_extrapolation = 0.1
)
}
\arguments{
\item{ts}{Time series in \code{s2ts} format (generated using \code{smooth_s2ts()}).}

\item{frequency}{(optional) One of the followings:
\itemize{
\item \code{daily}: daily frequency (default);
\item \code{dop} (Days Of Passage): values are returned corresponding to
the theoretic Sentinel-2 dates of passage.
}}

\item{method}{(optional) Argument passed to \code{spline()} or \code{approx()};
if \code{method = "hermite"}, \code{pracma::pchip()} is used.}

\item{max_na_days}{(optional) maximum number of consecutive days with missing
values which can be filled (in case of longer time windows with missing data,
NA are returned).
Default is to fit everything (unless this could lead to errors in case of
long NT time windows, currently it is the only way to get subsequent functions
working).}

\item{max_extrapolation}{(optional) Numeric: maximum allowed extrapolation
out of original range (relative value).
Default is 0.1 (+10\%). Set to Inf in order not to set any constraint.}
}
\value{
The output time series in tabular format (see \code{extract_ts()}).
}
\description{
Fill temporal gaps in a time series smoothed with function
\code{\link[=smooth_s2ts]{smooth_s2ts()}} to obtain homogeneous values (daily or with a regular
time step).
}
\examples{
#' # Load input data
data("ts_smoothed")

# Gap filling using default parameters (daily)
ts_filled <- fill_s2ts(ts_smoothed)
ts_filled # standard print
head(as.data.frame(ts_filled)) # see content
plot(ts_filled)

# Generate a regular time series using the minimum number of required records
ts_filled_2 <- fill_s2ts(ts_smoothed, frequency = "dop")
print(ts_filled_2, topn = 5) # standard print
}
\author{
Luigi Ranghetti, PhD (2020) \email{luigi@ranghetti.info}
}
