% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_s2ts.R
\name{fill_s2ts}
\alias{fill_s2ts}
\title{Fill not equally-spaced time series}
\usage{
fill_s2ts(ts, frequency = "dop", method = "fmm", max_na_days = 30)
}
\arguments{
\item{ts}{Time series in \code{s2ts} format (generated using \code{smooth_s2ts()}).}

\item{frequency}{(optional) One of the followings:
\itemize{
\item \code{dop} (Days Of Passage -- default): values are returned corresponding to
the theoretic Sentinel-2 dates of passage;
\item \code{daily}: daily frequency.
}}

\item{method}{(optional) Argument passed to \code{spline()}.}

\item{max_na_days}{(optional) maximum number of consecutive days with missing
values which can be filled (in case of longer time windows with missing data,
NA are returned).}
}
\value{
The output time series in tabular format (see \code{extract_ts()}).
}
\description{
TODO
}
\author{
Luigi Ranghetti, PhD (2020) \email{luigi@ranghetti.info}
}
