% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_curve.R
\name{fit_curve}
\alias{fit_curve}
\title{Fit curve over seasons}
\usage{
fit_curve(ts, seas, fit = "klosterman")
}
\arguments{
\item{ts}{Time series in \code{s2ts} format (generated using \code{fill_s2ts()}).}

\item{seas}{Seasons allocation (data table generated by \code{cut_seasons()}).}

\item{fit}{Fitting function among \code{"klosterman"}, \code{"beck"},
\code{"elmore"} and \code{"gu"} (see \code{phenopix::greenProcess()}).}
}
\value{
The output time series in tabular format (see \code{extract_ts()}),
with a new column \code{fit} containing fitted values.
}
\description{
Fit a curve using a parametric function  from
\code{phenopix::greenProcess()}.
}
\author{
Luigi Ranghetti, PhD (2020) \email{luigi@ranghetti.info}
}
