% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_sen2r.R
\name{print_message}
\alias{print_message}
\title{Print a message}
\usage{
print_message(
  ...,
  type,
  sep = "",
  date = FALSE,
  date_format = "",
  width = 0.9 * getOption("width"),
  indent = TRUE,
  exdent = TRUE,
  prefix = "",
  initial = prefix
)
}
\arguments{
\item{...}{\code{R} objects which are concatenated.}

\item{type}{Type of the output. Accepted values: \code{'message'}, \code{'string'},
\code{'cat'}, \code{'error'}, \code{'warning'}, \code{'waiting'}.}

\item{sep}{(optional) character used to separate input values.}

\item{date}{Logical: set \code{TRUE} to place the date before the message.}

\item{date_format}{Format of the date (see \code{strftime()}).}

\item{width}{Positive integer: target column for wrapping lines.}

\item{indent, exdent}{Indentation of the first / subsequent lines.}

\item{prefix, initial}{Prefix for each line / for the first line.}
}
\value{
Message (in the defined format).
}
\description{
A common interface for printing messages of several types.
}
\note{
Originally from package sen2r (GPL-3).
}
\author{
Luigi Ranghetti, PhD (2020)
}
\keyword{internal}
