% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_s2cube.R
\name{read_s2cube}
\alias{read_s2cube}
\title{Read raster cubes from sen2r archives}
\usage{
read_s2cube(
  inpath,
  out_format = "path",
  prod_type,
  time_window,
  s2_orbits,
  s2_sensors,
  file_ext,
  bbox,
  .use_vrt = FALSE
)
}
\arguments{
\item{inpath}{Path of the directory in which sen2r files are.}

\item{out_format}{(optional) Output format, being one among
\code{"path"} (default), \code{"stars"} and \code{"stars_proxy"}).}

\item{prod_type}{(optional) sen2r product type to import
(used for filtering among \code{inpath} content).}

\item{time_window}{(optional) time window to import
(used for filtering among \code{inpath} content).}

\item{s2_orbits}{(optional) Sentinel-2 orbits to import
(used for filtering among \code{inpath} content).}

\item{s2_sensors}{(optional) Sentinel-2 sensor to import  (\code{"2A"}, \code{"2B"} or both)
(used for filtering among \code{inpath} content).}

\item{file_ext}{(optional) input file extension
(used for filtering among \code{inpath} content).}

\item{bbox}{(optional) object of class \code{sf}, \code{sfc} or \code{bbox} used to import
a subset of the original raster extent.}

\item{.use_vrt}{(optional) pass through a temporary VRT before creating \code{stars}?}
}
\value{
The output raster cube in the selected format.
}
\description{
TODO
}
\examples{
# Path of a sample archive created with {sen2r}
archive_dir <- system.file("extdata/sen2r/mincrops", package = "sen2rts")
# Default behaviour
sen2r_ndvi_paths <- read_s2cube(file.path(archive_dir, "NDVI"))
# Specifying argument prod_type, the parent directory can also be provided:
sen2r_ndvi_paths <- read_s2cube(archive_dir, prod_type = "NDVI")
head(sen2r_ndvi_paths)
# Using some filters
read_s2cube(
  archive_dir, prod_type = "NDVI", 
  time_window = c("2020-06-01", "2020-06-30"), 
  s2_sensors = "2A"
)
}
\author{
Luigi Ranghetti, PhD (2020) \email{luigi@ranghetti.info}
}
