% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2_dop_simpl.R
\name{s2_dop_simpl}
\alias{s2_dop_simpl}
\title{Simplified \code{sen2r::s2_dop()} for internal use}
\usage{
s2_dop_simpl(s2_orbits, timewindow, mission)
}
\arguments{
\item{s2_orbits}{A vector of Sentinel-2 orbits (as integer numbers
or 3-length character).}

\item{timewindow}{Temporal window for querying: Date object
of length 2 (time window).}

\item{mission}{(optional) Vector with the desired Sentinel-2 missions
("2A", "2B" or both).}
}
\value{
A data table with the dates (column "date"), the missions
(column "mission") and the orbits (column "orbit").
}
\description{
This is a simplified version of \code{s2_dop()} which returns
Sentinel-2 passages over certain orbits during a defined time interval.
Checks over inputs are skipped to save time, and S2B dates before
Sentinel-2 launch can also be returned, since this is required by
\code{smooth_ts()}.
}
\note{
License: GPL 3.0
}
\author{
Luigi Ranghetti, PhD (2021) \email{luigi@ranghetti.info}
}
