% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_ts.R
\name{smooth_s2ts}
\alias{smooth_s2ts}
\title{Filter and smooth time series from sen2r archives}
\usage{
smooth_s2ts(
  ts,
  min_qa = 0.5,
  noise_dir = "undefined",
  spike = 0.25,
  spike_window = 3,
  sg_window = 9,
  sg_polynom = 3,
  keep_max = FALSE
)
}
\arguments{
\item{ts}{Time series in \code{s2ts} format (generated using \code{extract_s2ts()}).}

\item{min_qa}{(optional) minimum 0-1 quality value
(points with \code{qa < min_qa} are not used, while \code{qa} values
in the range \code{min_qa} to 1 are reshaped ito the range 0 to 1).
Default is 0.5.}

\item{noise_dir}{Direction of points generally containing noise.
If \code{"low"}, higher values are generally maintained (this is the case of
most vegetation indices like NDVI);
if \code{"high"}, lower values are generally maintained;
if \code{"undefined"} (default), no assumptions are done.}

\item{spike}{Relative "y" difference for spike determination (default is 0.25).}

\item{spike_window}{Maximum number of values for spike identification
(it must be an odd number). Default is 3.}

\item{sg_window}{Argument \code{window} of function \code{w_savgol()}.}

\item{sg_polynom}{Argument \code{polynom} of function \code{w_savgol()}.}

\item{keep_max}{(optional) Logical: if TRUE, the maximum (if \code{noise_dir = "low"})
or minimum (if \code{noise_dir = "high"}) value between the input and the smoothed
is finally keeped; if FALSE (default) smoothed values are always keeped.
This argument is not used if \code{noise_dir = "undefined"}.}
}
\value{
The output time series in \code{s2ts} format.
}
\description{
TODO
}
\author{
Luigi Ranghetti, phD (2020) \email{luigi@ranghetti.info}
}
