% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_s2ts.R
\name{smooth_s2ts}
\alias{smooth_s2ts}
\title{Filter and smooth time series from sen2r archives}
\usage{
smooth_s2ts(
  ts,
  min_qa = 0.2,
  noise_dir = "low",
  spike = 0.25,
  spike_window = 5,
  sg_daywindow = 15,
  sg_polynom = 2,
  sg_n = 3,
  max_extrapolation = 0.1
)
}
\arguments{
\item{ts}{Time series in \code{s2ts} format (generated using \code{extract_s2ts()}).}

\item{min_qa}{(optional) minimum 0-1 quality value
(points with \code{qa < min_qa} are not used, while \code{qa} values
in the range \code{min_qa} to 1 are reshaped into the range 0 to 1).
Default is 0.5.}

\item{noise_dir}{Direction of points generally containing noise.
If \code{"low"}, higher values are generally maintained (this is the case of
most vegetation indices like NDVI);
if \code{"high"}, lower values are generally maintained;
if \code{"undefined"} (default), no assumptions are done.}

\item{spike}{Relative "y" difference for spike determination (default is 0.25).
Set to NA to skip spike removal.}

\item{spike_window}{Maximum number of values for spike identification
(it must be an odd number). Default is 3.}

\item{sg_daywindow}{Half-size of the time window to be used to interpolate}

\item{sg_polynom}{Argument \code{polynom} of function \code{w_savgol()}.}

\item{sg_n}{(optional) Positive integer: number of applications of the
Savitzky-Golay filter. The minimum value is 1 (a single application
using the weights included in \code{ts}).
Within each additional application, the weights included in \code{ts} are
multiplicated for the relative ranks of the difference between original
and smoothed values (if \code{noise_dir = "low"}) or between smoothed and
original values (if \code{noise_dir = "high"}).
If \code{noise_dir = "undefined"}, this argument is coerced to 1.}

\item{max_extrapolation}{(optional) Numeric: maximum allowed extrapolation
out of original range (relative value).
Default is 0.1 (+10\%). Set to Inf in order not to set any constraint.}
}
\value{
The output time series in \code{s2ts} format.
}
\description{
Smooth time series extracted with function \code{\link[=extract_s2ts]{extract_s2ts()}}
using a Savitzky-Golay filter.
Quality flags associated to the time series are used to obtain better
results.
Argument values should be manually set to obtain a correct smoothing.
}
\examples{
# Load input data
data("ts_raw")

# Smoothing time series using default parameters
ts_smoothed <- smooth_s2ts(ts_raw)
print(ts_smoothed, topn = 5) # standard print
head(as.data.frame(ts_smoothed)) # see content
plot(ts_smoothed)

# Apply a more pronounced smoothing
ts_smoothed_2 <- smooth_s2ts(
  ts_raw,
  min_qa = 0.5, # exclude values with qa < 0.5
  sg_daywindow = 30, # larger moving window
  sg_polynom = 3, # cubic interpolation
  sg_n = 5 # apply the SG filter 5 times
)
plot(ts_smoothed_2)
}
\author{
Luigi Ranghetti, PhD (2020) \email{luigi@ranghetti.info}
}
