% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_poi.R
\name{st_poi}
\alias{st_poi}
\title{Extract phenological metrics}
\usage{
st_poi(x, ..., precision)
}
\arguments{
\item{x}{object of class \code{sfg}, \code{sfc} or \code{sf}}

\item{...}{Ignored}

\item{precision}{The precision to use when computing the centre
(passed to \code{polylabelr::poi()}), in the \code{x} unit.
If missing, by default it is used the 1/1000 of the mean square root of the
area of inputs.}
}
\description{
This function computes and returns the approximate pole of
inaccessibility (visual centre) for a polygon using the quadtree-based
algorithm implemented in \code{polylabelr::poi()},
accepting an input in the same format of \code{sf_centroid()}
(\code{sf}, \code{sfc} or \code{sfg})
and returning an output in the same format of \code{x}.
}
\author{
Luigi Ranghetti, PhD (2021) \email{luigi@ranghetti.info}
}
