% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_warp_fixing.R
\name{st_warp_fixing}
\alias{st_warp_fixing}
\title{Extract phenological metrics}
\usage{
st_warp_fixing(src, dest, ...)
}
\arguments{
\item{src}{object of class \code{stars} with source raster}

\item{dest}{object of class \code{stars} with target raster geometry}

\item{...}{Additional arguments passed to \code{stars::st_warp()}.}
}
\description{
Internal function to bypass an encountered error:
sometimes (apparently with large stacks) returns stars_proxy
instead than stars: in this case, manually build a stars object.
Temporary file paths created on source (this occurs reading VRT) are
stored in the attribute "tempsourcepath".
}
\author{
Luigi Ranghetti, PhD (2020) \email{luigi@ranghetti.info}
}
