% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w_savgol.R
\name{w_savgol}
\alias{w_savgol}
\title{Savitzky-Golay filter for not equally-spaced weighted data}
\usage{
w_savgol(y, x, q, window = 7, polynom = 3)
}
\arguments{
\item{y}{List of floats representing the "y" values.}

\item{x}{List of floats representing the "x" values of the data.
Must have same length as \code{y}.
If missing, points are assumed to be equally-spaced}

\item{q}{List of floats representing the relative weight of "y" values.
Must have same length as \code{y}.
If missing, all points are assumed to have the same weight.}

\item{window}{Window length of datapoints.
Must be odd and smaller than \code{x}. Default is 7.}

\item{polynom}{The polynomial order to be used.
Must be smaller than the \code{window} size. Default is 3.}
}
\value{
The smoothed "y" values.
}
\description{
Smooth values using a Savitzky-Golay filter.
This function is suitable for not equally-spaced and/or weighted data.
}
\note{
This is an R adaptation of Python function at
\url{https://dsp.stackexchange.com/questions/1676/savitzky-golay-smoothing-filter-for-not-equally-spaced-data},
with the addition of weights following
\url{https://en.wikipedia.org/wiki/Savitzky-Golay_filter}.
}
\author{
Luigi Ranghetti, PhD (2020) \email{luigi@ranghetti.info}
}
